contract/requirements.txt
# Core Dependencies
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# NLP & ML
spacy>=3.7.0
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.7.0
chromadb>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Translation & Language
deep-translator>=1.11.0
googletrans>=4.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Export & Reporting
reportlab>=4.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Security
cryptography>=41.0.0
bcrypt>=4.1.0
python-jose>=3.3.0
passlib>=1.7.4

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
